Refactor the TimePathSource to use extractor/vertractor
